landing and auth pages perf optimize
