+++ src/app/(auth)/register/page.tsx
@@ -1,99 +1,12 @@
-"use client";
-
-import { useState } from "react";
-import { useAuth } from "@/app/context/AuthContext";
 import Link from "next/link";
 import Image from "next/image";
-import { Eye, EyeOff, Loader2 } from "lucide-react";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import { getApiErrorMessage } from "@/types";
+import { RegisterForm } from "@/components/auth/RegisterForm";
 
-export default function Register() {
-  const [formData, setFormData] = useState({
-    firstName: "",
-    lastName: "",
-    email: "",
-    phone: "",
-    password: "",
-    confirmPassword: "",
-  });
-  const [error, setError] = useState("");
-  const [showPassword, setShowPassword] = useState(false);
-  const [isLoading, setIsLoading] = useState(false);
-
-  const { register } = useAuth();
-
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setFormData({ ...formData, [e.target.name]: e.target.value });
-  };
-
-  const passwordChecks = [
-    { label: "8+ characters", met: formData.password.length >= 8 },
-    { label: "Uppercase letter", met: /[A-Z]/.test(formData.password) },
-    { label: "Lowercase letter", met: /[a-z]/.test(formData.password) },
-    { label: "Number", met: /\d/.test(formData.password) },
-  ];
-  const passedCount = passwordChecks.filter((c) => c.met).length;
-  const strengthLabel =
-    formData.password.length === 0
-      ? ""
-      : passedCount <= 1
-        ? "Weak"
-        : passedCount === 2
-          ? "Fair"
-          : passedCount === 3
-            ? "Good"
-            : "Strong";
-  const strengthColor =
-    passedCount <= 1
-      ? "bg-red-500"
-      : passedCount === 2
-        ? "bg-amber-500"
-        : passedCount === 3
-          ? "bg-blue-500"
-          : "bg-emerald-500";
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setError("");
-
-    if (passedCount < 4) {
-      setError(
-        "Password must be at least 8 characters with uppercase, lowercase, and a number",
-      );
-      return;
-    }
-
-    if (formData.password !== formData.confirmPassword) {
-      setError("Passwords do not match");
-      return;
-    }
-
-    setIsLoading(true);
-    try {
-      // Pass as object to match new AuthContext signature
-      await register({
-        firstName: formData.firstName,
-        lastName: formData.lastName,
-        email: formData.email,
-        phone: formData.phone,
-        password: formData.password,
-        confirmPassword: formData.confirmPassword,
-      });
-    } catch (err: unknown) {
-      setError(getApiErrorMessage(err, "Registration failed"));
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
+export default function RegisterPage() {
   return (
     <div className="flex min-h-screen w-full bg-slate-50 font-sans">
       {/* Left Side - Navy Theme */}
       <div className="hidden lg:flex w-1/2 bg-[var(--navy)] relative flex-col p-16 text-white overflow-hidden">
-        {/* Background Curves */}
         <div className="absolute inset-0 opacity-10 pointer-events-none">
           <svg
             className="h-full w-full"
@@ -122,7 +35,6 @@ export default function Register() {
             <h1 className="text-5xl font-bold leading-tight mb-5">
               Create your account.
             </h1>
-
             <p className="text-slate-400 text-lg xl:text-xl max-w-lg leading-relaxed">
               Join the platform managing complex construction logistics. Explore
               key workflows and see the efficiency gains.
@@ -137,161 +49,7 @@ export default function Register() {
 
       {/* Right Side - Register Form */}
       <div className="w-full lg:w-1/2 flex items-center justify-center p-6 sm:p-12 overflow-y-auto">
-        <div className="w-full max-w-md space-y-8 my-auto">
-          <div className="text-center lg:text-left">
-            <h2 className="text-3xl font-bold text-slate-900">
-              Create Account
-            </h2>
-            <p className="text-slate-500 mt-2">Get started with SiteSpace</p>
-          </div>
-
-          {error && (
-            <div
-              className="bg-red-50 border border-red-100 text-red-600 px-4 py-3 rounded-lg text-sm font-medium animate-in slide-in-from-top-2"
-              role="alert"
-            >
-              {error}
-            </div>
-          )}
-
-          <form onSubmit={handleSubmit} className="space-y-4">
-            <div className="grid grid-cols-1 gap-4 sm:grid-cols-2">
-              <div className="space-y-2">
-                <Label htmlFor="firstName">First Name</Label>
-                <Input
-                  id="firstName"
-                  name="firstName"
-                  placeholder="John"
-                  onChange={handleChange}
-                  required
-                  className="h-11 border-slate-200 focus-visible:ring-[var(--navy)]"
-                />
-              </div>
-              <div className="space-y-2">
-                <Label htmlFor="lastName">Last Name</Label>
-                <Input
-                  id="lastName"
-                  name="lastName"
-                  placeholder="Doe"
-                  onChange={handleChange}
-                  required
-                  className="h-11 border-slate-200 focus-visible:ring-[var(--navy)]"
-                />
-              </div>
-            </div>
-
-            <div className="space-y-2">
-              <Label htmlFor="email">Email</Label>
-              <Input
-                id="email"
-                name="email"
-                type="email"
-                placeholder="name@company.com"
-                onChange={handleChange}
-                required
-                className="h-11 border-slate-200 focus-visible:ring-[var(--navy)]"
-              />
-            </div>
-
-            <div className="space-y-2">
-              <Label htmlFor="phone">Phone</Label>
-              <Input
-                id="phone"
-                name="phone"
-                type="tel"
-                placeholder="+1 (555) 000-0000"
-                onChange={handleChange}
-                required
-                className="h-11 border-slate-200 focus-visible:ring-[var(--navy)]"
-              />
-            </div>
-
-            <div className="space-y-2">
-              <Label htmlFor="password">Password</Label>
-              <div className="relative">
-                <Input
-                  id="password"
-                  name="password"
-                  type={showPassword ? "text" : "password"}
-                  placeholder="••••••••"
-                  onChange={handleChange}
-                  required
-                  className="h-11 pr-10 border-slate-200 focus-visible:ring-[var(--navy)]"
-                />
-                <button
-                  type="button"
-                  onClick={() => setShowPassword(!showPassword)}
-                  className="absolute right-3 top-1/2 -translate-y-1/2 text-slate-400 hover:text-slate-600"
-                  aria-label={showPassword ? "Hide password" : "Show password"}
-                >
-                  {showPassword ? <EyeOff size={18} /> : <Eye size={18} />}
-                </button>
-              </div>
-              {formData.password.length > 0 && (
-                <div className="space-y-2 mt-2">
-                  <div className="flex items-center gap-2">
-                    <div className="flex-1 h-1.5 bg-slate-100 rounded-full overflow-hidden">
-                      <div
-                        className={`h-full rounded-full transition-all duration-300 ${strengthColor}`}
-                        style={{ width: `${(passedCount / 4) * 100}%` }}
-                      />
-                    </div>
-                    <span
-                      className={`text-xs font-medium ${passedCount <= 1 ? "text-red-600" : passedCount === 2 ? "text-amber-600" : passedCount === 3 ? "text-blue-600" : "text-emerald-600"}`}
-                    >
-                      {strengthLabel}
-                    </span>
-                  </div>
-                  <div className="flex flex-wrap gap-x-3 gap-y-1">
-                    {passwordChecks.map((check) => (
-                      <span
-                        key={check.label}
-                        className={`text-[11px] ${check.met ? "text-emerald-600" : "text-slate-400"}`}
-                      >
-                        {check.met ? "\u2713" : "\u2022"} {check.label}
-                      </span>
-                    ))}
-                  </div>
-                </div>
-              )}
-            </div>
-
-            <div className="space-y-2">
-              <Label htmlFor="confirmPassword">Confirm Password</Label>
-              <Input
-                id="confirmPassword"
-                name="confirmPassword"
-                type="password"
-                placeholder="••••••••"
-                onChange={handleChange}
-                required
-                className="h-11 border-slate-200 focus-visible:ring-[var(--navy)]"
-              />
-            </div>
-
-            <Button
-              type="submit"
-              className="w-full h-12 bg-[var(--navy)] hover:bg-[var(--navy-hover)] text-white font-bold text-base transition-all mt-4"
-              disabled={isLoading}
-            >
-              {isLoading ? (
-                <Loader2 className="animate-spin h-5 w-5" />
-              ) : (
-                "Create Account"
-              )}
-            </Button>
-          </form>
-
-          <p className="text-center text-sm text-slate-500">
-            Already have an account?{" "}
-            <Link
-              href="/login"
-              className="font-bold text-[var(--navy)] hover:underline"
-            >
-              Log in
-            </Link>
-          </p>
-        </div>
+        <RegisterForm />
       </div>
     </div>
   );
