+++ src/app/context/AuthContext.tsx
@@ -341,18 +341,6 @@ export function AuthProvider({ children }: { children: ReactNode }) {
     [user, isInitialized, error, login, register, logout, clearError],
   );
 
-  if (!isInitialized) {
-    return (
-      <div className="min-h-screen flex items-center justify-center bg-slate-50">
-        <div
-          className="animate-spin h-8 w-8 border-4 border-slate-200 border-t-slate-800 rounded-full"
-          role="status"
-          aria-label="Loading"
-        />
-      </div>
-    );
-  }
-
   return <AuthContext.Provider value={value}>{children}</AuthContext.Provider>;
 }
